+++ tests/data/CellAbsoluteReference.php
@@ -26,12 +26,12 @@ return [
         'AI2012',
     ],
     [
-        '\'Worksheet1\'!$AI$256',
-        '\'Worksheet1\'!AI256',
+        '\'Work!sheet1\'!$AI$256',
+        '\'Work!sheet1\'!AI256',
     ],
     [
-        'Worksheet1!$AI$256',
-        'Worksheet1!AI256',
+        'Work!sheet1!$AI$256',
+        'Work!sheet1!AI256',
     ],
     [
         '\'Data Worksheet\'!$AI$256',